requirements.txt
django==5.1.1
djangorestframework==3.15.2
python-dotenv==1.0.1
Pillow==10.4.0
celery==5.4.0
redis==5.0.8